// src/app/client-view/models/current-scenario/cash-flow-details.model.ts
export class CashFlowDetails {
    deductibleDebt: number;
    governmentBenefits: number;
    grossSalary: number;
    incomeTax: number;
    investmentContributions: number;
    investmentExpenses: number;
    investmentIncome: number;
    lifestyleGoals: number;
    lifestyleNeeds: number;
    lifestyleWants: number;
    nonDeductibleDebt: number;
    otherIncome: number;
    period: {
        frequency: number;
        value: number;
    };
    personalInsurance: number;
    retirementIncomeStreams: number;
    preTaxSpending: number;
    salarySacrifice: number;
    surplusIncome: number;
    totalExpenses: number;
    totalGrossIncome: number;
    totalNetIncome: number;
    weeklySurplus: number;
    netAccountTransfers: number;
}